+++ src/Events/QueuedDeleteAll.php
@@ -2,7 +2,39 @@
 
 namespace Ikechukwukalu\Clamavfileupload\Events;
 
-class QueuedDeleteAll extends QueuedDelete
-{
+use Illuminate\Broadcasting\Channel;
+use Illuminate\Broadcasting\InteractsWithSockets;
+use Illuminate\Broadcasting\PresenceChannel;
+use Illuminate\Broadcasting\PrivateChannel;
+use Illuminate\Contracts\Broadcasting\ShouldBroadcast;
+use Illuminate\Foundation\Events\Dispatchable;
+use Illuminate\Queue\SerializesModels;
 
+class QueuedDeleteAll
+{
+    use Dispatchable, InteractsWithSockets, SerializesModels;
+
+    public string $ref;
+
+    /**
+     * Create a new event instance.
+     *
+     * @param  string  $ref
+     */
+    public function __construct(string $ref)
+    {
+        $this->ref = $ref;
+    }
+
+    /**
+     * Get the channels the event should broadcast on.
+     *
+     * @return array<int, \Illuminate\Broadcasting\Channel>
+     */
+    public function broadcastOn(): array
+    {
+        return [
+            new PrivateChannel('channel-name'),
+        ];
+    }
 }