+++ src/Events/QueuedForceDeleteAll.php
@@ -2,7 +2,7 @@
 
 namespace Ikechukwukalu\Clamavfileupload\Events;
 
-class QueuedForceDeleteAll extends QueuedDelete
+class QueuedForceDeleteAll extends QueuedDeleteAll
 {
 
 }